s3-build: Remove build of libaddns.a from the autoconf build
We always link libaddns statically as part of the net object lists

This means that we no longer provide externally-available
libaddns, as the waf build declared this as a private library.

This never had a public API, .so or a header file.

Andrew Bartlett

Autobuild-User: Andrew Bartlett <abartlet@samba.org>
Autobuild-Date: Sat May 26 09:45:34 CEST 2012 on sn-devel-104