libreplace: Add copyrights to ease tracking of this file in future
As GIT didn't realise this was a copy out of lib/system.c, this should
make it easier to track the copyright holders on this file.  Herb's
name wasn't on the original file, but was the only other author I
could find in the git logs.

I've added my copyright here too.

Andrew Bartlett