regedit: Setup a test tree view of HKLM.
Expand on the former test by allowing the user to walk HKLM
with the tree view. Subkeys are loaded dynamically in memory
as the user walks the hierarchy with tree_node_load_children().

One issue with this at the moment is detecting if the
in-memory copy of the hierarchy is out of date.

TODO: figure out when to reload out of date portions of the
hierarchy (probably by checking last_modified_time).

Reviewed-by: Andreas Schneider <asn@samba.org>
Reviewed-by: Michael Adam <obnox@samba.org>